server/requirements.txt
Django==4.1.5

django-environ==0.9.0
django-extensions===3.2.1
django-cors-headers==3.13.0

djangorestframework==3.14.0
django-filter==22.1
